Updated implementation to be more compatible with how I'd use it
Added SimpleBuilder interface which is really just a rename of Builder so there are no naming conflicts for implementations also named "Builder"
AbstractBuilder takes an additional type argument for the "constructor" object, or object used to store the mutable state while constructing
AbstractBuilder now implements the Function interface so that a translation function is defined to translate from the constructor object instance to the object that is being built (this is implemented to be called automatically)
AbstractBuilder now captures the current state of the constructing object instead of returning a static reference (this can still be done in a subclass though)
Builder is no longer a Supplier of the generic class. This was not needed, and only existed because of the erroneous implementation within AbstractBuilder to expect an instance of the object being built to store a mutable state, however this is not the case with builders for immutable instances
Updated documentation to more accurately reflect new and old function, however I still want to add a package-info.java to explain the use-case of this, as well as it's deficiencies, because it's not perfect because of language limitations.
Modified BuildableImpl and BuildableLinkedList to accommodate the above changes